GRSPW2: 	minor change to auto_drop api 	add auto_drop call to libc

+++ include/grspw2.h
@@ -632,7 +632,7 @@ struct grspw2_data {
 	void *data;
 	uint32_t data_size;
 	uint8_t *pkt;
-	int n_drop;
+	uint8_t n_drop;
 };
